+++ assets/javascripts/kitten/components/layout/dashboard-layout/stories.js
@@ -71,7 +71,6 @@ const options = [
   },
 ]
 
-
 import { Default as Table } from '../../tables/list-table/list-table.stories.js'
 
 import { useWindowWidth } from '../../../helpers/utils/use-window-width-hook'
@@ -313,10 +312,15 @@ export const Default = () => {
         </Button>
       </DashboardLayout.SideFooter>
 
-      {boolean('Display Alert', false) && (
-        <DashboardLayout.Alert closeButton error>
-          Voilà une alerte
-        </DashboardLayout.Alert>
+      {boolean('Display Alerts', false) && (
+        <DashboardLayout.Alerts>
+          <Alert closeButton info>
+            Voilà une info
+          </Alert>
+          <Alert closeButton error>
+            Voilà une alerte
+          </Alert>
+        </DashboardLayout.Alerts>
       )}
 
       {selectedView === 'flow' && <FlowExample />}
@@ -379,14 +383,14 @@ const FlowExample = () => (
         augue laoreet rutrum faucibus dolor auctor.
       </p>
       <DropdownSelectWithInput
-          id="DropdownSelectWithInput"
-          hideLabel
-          labelText="Facebook"
-          options={options}
-          highlightOptionBox
-          defaultSelectedValue="facebook"
-          deactivateDropdown
-          className="k-u-margin-bottom-triple"
+        id="DropdownSelectWithInput"
+        hideLabel
+        labelText="Facebook"
+        options={options}
+        highlightOptionBox
+        defaultSelectedValue="facebook"
+        deactivateDropdown
+        className="k-u-margin-bottom-triple"
       />
       <p className="k-u-weight-light">
         Lorem ipsum dolor sit amet, consectetur adipiscing elit. Duis mollis,